adding latest production build
