+++ crawlerapp/public/assets/components/d3-visualization-cda85a4657c912e5b96dcd787d4c708a.css
@@ -1,17 +1,19 @@
 .backend-list {
-  background: white;
-  box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);
+  background: transparent;
+  /* box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23); */
   padding: 1%;
+  margin-top: 50px
 }
 
 .links line {
     stroke: #999;
-    stroke-opacity: 0.6;
+    pointer-events: all;
   }
-   
+
   .nodes circle {
     stroke: #ea1a1a;
     stroke-width: 1.5px;
+    pointer-events: all;
   }
 
   text {
@@ -22,7 +24,7 @@
   div.tooltip {
     position: absolute;
     background-color: white;
-    max-width: 250px;
+    max-width: 350px;
     height: auto;
     padding: 1px;
     border-style: solid;
@@ -30,4 +32,10 @@
     border-width: 1px;
     box-shadow: 3px 3px 10px rgba(0, 0, 0, .5);
     pointer-events: none;
+    padding: 7px;
+  }
+
+  .tooltipLabel{
+    text-decoration: underline;
+    font-weight: bold;
   }
