+++ crawlerapp/public/assets/components/nav-bar-1b4716f7223a35426b3fd591ba601e1e.css
@@ -1,6 +1,11 @@
 .nav-wrapper {
-    background: rgb(48,190,162);
-    background: linear-gradient(90deg, rgba(48,190,162,0.8547794117647058) 0%, rgba(109,107,159,1) 100%, rgba(44,105,88,0.7791491596638656) 100%);
+    background: #0F2027;  /* fallback for old browsers */
+    background: -webkit-linear-gradient(to right, #2C5364, #203A43, #0F2027);  /* Chrome 10-25, Safari 5.1-6 */
+    background: linear-gradient(to right, #2C5364, #203A43, #0F2027); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
+}
+
+.brand-logo {
+    margin-left: 0px;
 }
 
 nav .brand-logo {
