+++ crawlerapp/public/assets/search-1092c6b314fddbb6002009c73fa83a53.css
@@ -1,0 +1,103 @@
+.search-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+}
+
+.search-container .history-header {
+    background-color:rgba(0, 23, 31, 0.774);
+    color: white;
+ }
+ 
+ .search-container .history-more-info {
+    background-color:rgba(0, 23, 31, 0.774);
+    color: white;
+}
+ 
+
+.search-button {
+    margin: 50px 0;
+    flex-basis: 20%;
+    text-align: center;
+}
+
+.search-visualization {
+    margin-bottom: 500px;
+}
+
+.search-error {
+    margin: 100px 0 500px 0;
+    color: white;
+    text-align: center;
+}
+
+.return-button .btn {
+        background-color: transparent;
+        border: 1px solid #007EA7;;
+        color: #00171f;;
+}
+
+.return-button .btn:hover {
+    background-color:#00A8E8;
+    color: white;
+}
+
+.node-key {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    flex-basis: 20%;
+    height: 100px;
+    padding: 1%;
+    background: transparent;
+    color: #00171f;
+    border: 1px solid #007EA7;
+    box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);
+}
+
+.node-item {
+    display: flex;
+    justify-content: space-between
+}
+.default-node {
+    background-color: #618BEC;
+    border: 1px solid red;
+    align-self: center;
+    width: 20px;
+    height: 20px;
+    border-radius: 50%;
+}
+
+.source-node {
+    background-color: #72FA77;
+    border: 1px solid red;
+    align-self: center;
+    width: 20px;
+    height: 20px;
+    border-radius: 50%; 
+}
+
+.keyword-node {
+    background-color: #F1FA72;
+    border: 1px solid red;
+    align-self: center;
+    width: 20px;
+    height: 20px;
+    border-radius: 50%; 
+}
+
+.search-row-container {
+    width: 1500px;
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    margin-top: 50px;
+    word-wrap: break-word;
+}
+
+.search-info-container-visualization {
+    background-color: white;
+    flex-basis: 30%;
+    box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);
+}
