--- crawlerapp/public/assets/search-72fc524a056d08a368ded9fc98b1222b.css
@@ -1,10 +1,0 @@
-.search-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
-}
-
-.search-button {
-    margin: 50px 0;
-}
