+++ crawlerapp/public/index.html
@@ -3,12 +3,12 @@
   <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Frontend</title>
+    <title>Team Azha Web Crawler</title>
     <meta name="description" content="">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
     
-<meta name="frontend/config/environment" content="%7B%22modulePrefix%22%3A%22frontend%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22%2F%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%7D%2C%22APP%22%3A%7B%22name%22%3A%22frontend%22%2C%22version%22%3A%220.0.0%2B62515ddd%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
+<meta name="frontend/config/environment" content="%7B%22modulePrefix%22%3A%22frontend%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22%2F%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%7D%2C%22APP%22%3A%7B%22name%22%3A%22frontend%22%2C%22version%22%3A%220.0.0%2Bcc2bfe51%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
     <!-- Compiled and minified CSS -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
     
@@ -18,7 +18,7 @@
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Abel|Allan&display=swap" rel="stylesheet">
     <link integrity="" rel="stylesheet" href="/assets/vendor-6958e8610e934b23f7a7f00c1137622b.css">
-    <link integrity="" rel="stylesheet" href="/assets/frontend-c81955c129cb1e04bea09ffe77cf1287.css">
+    <link integrity="" rel="stylesheet" href="/assets/frontend-60ba5b3fa197811c589431f70ba5cc54.css">
 
     
   </head>
@@ -26,7 +26,7 @@
     
 
     <script src="/assets/vendor-4c1720f8d96276969fe2f55426268ff2.js" integrity="sha256-bFy0q7i7gplTmQyxTE8G+6FdJ8mqH/g4/xQo0Imed78= sha512-ks+f3EV6gvH+KY/TFZc8VJCGsOmMAPeFLmpnBoZFyBadhLgxeXx7x8Un/pD4bydABB2siuDCwd9xrZqRLaF3Ew==" ></script>
-    <script src="/assets/frontend-0b6f00b646ba57f51eada978dfb96cce.js" integrity="sha256-0sWfti+mjaehM5+eJIPj8DP0uagvRSqwO8Th6Muu2VM= sha512-4nea+L6EsiAIfvO6rJWBlDY9QoVRTeQPbIeTlb72XkcDLhkkf/ttuSJffICNGaoAPnSdYFWXuyudkRueR3p6VA==" ></script>
+    <script src="/assets/frontend-4721b780da305fd5dcc92a93ad9d7bc6.js" integrity="sha256-i4/tkR951U8GNOD1omlYhnKK/6gcji+VWSecreSC3LI= sha512-AdzKpHnrmsnGtW4L5b5acbc9Y6ZvGSR5J9/dA+Z561f0cixWt9wJW9is1s2ZJYwA2AtMSV7QQVqGUUVPsPWwHg==" ></script>
 
     <!-- Compiled and minified JavaScript -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
